(visibly) removed background on touchscreen - colors don't work anymore, though
The background of this plugin is a little darker on the touchscreen after loading (and not switching devises; in "default" mode). I wanted to remove this as this part of the screen appeared highlighted all the time by this. 
Commenting out the opacity (or setting it to 0) made the text invisible. Thus, I set it to this really low value so the text is still visible.

This is only a dirty "fix" though. Now, switching devices does not show the colors (blue, red, green) anymore. For me this does not matter, but overall there surely is a smarter setting to change somewhere, I just couldn't find it. I played around with many variables that set the background color but I could not figure it out quickly. I'm just a hobby scripter :)

I hope you can maybe disable this somewhere else so the useful colored indicators when switching devices do not get destroyed.